+++ src/rte/RTESolver.jl
@@ -70,13 +70,14 @@ Additionally, takes an optional argument `metric_scaling` which scales the resul
 corresponding factor to account for column expansion in `deep-atmosphere` configurations.
 """
 function solve_lw!(
-    (; context, fluxb, flux, src, bcs, op, angle_disc)::NoScatLWRTE,
+    (; context, fluxb, flux, src, bcs, op, angle_disc, state_cache)::NoScatLWRTE,
     as::AtmosphericState,
     lookup_lw::LookUpLW,
     lookup_lw_cld::Union{LookUpCld, Nothing} = nothing,
     lookup_lw_aero::Union{LookUpAerosolMerra, Nothing} = nothing,
     metric_scaling::M = nothing,
 ) where {M}
+    AtmosphericStates.refresh_transposed_state!(state_cache, as)
     rte_lw_noscat_solve!(
         context.device,
         fluxb,
@@ -86,6 +87,7 @@ function solve_lw!(
         op,
         angle_disc,
         as,
+        state_cache,
         lookup_lw,
         lookup_lw_cld,
         lookup_lw_aero,
@@ -108,13 +110,14 @@ Additionally, takes an optional argument `metric_scaling` which scales the resul
 corresponding factor to account for column expansion in `deep-atmosphere` configurations.
 """
 function solve_lw!(
-    (; context, fluxb, flux, band_flux, src, bcs, op)::TwoStreamLWRTE,
+    (; context, fluxb, flux, band_flux, src, bcs, op, state_cache)::TwoStreamLWRTE,
     as::AtmosphericState,
     lookup_lw::LookUpLW,
     lookup_lw_cld::Union{LookUpCld, Nothing} = nothing,
     lookup_lw_aero::Union{LookUpAerosolMerra, Nothing} = nothing,
     metric_scaling::M = nothing,
 ) where {M}
+    AtmosphericStates.refresh_transposed_state!(state_cache, as)
     rte_lw_2stream_solve!(
         context.device,
         fluxb,
@@ -124,6 +127,7 @@ function solve_lw!(
         bcs,
         op,
         as,
+        state_cache,
         lookup_lw,
         lookup_lw_cld,
         lookup_lw_aero,
@@ -177,12 +181,22 @@ Additionally, takes an optional argument `metric_scaling` which scales the resul
 corresponding factor to account for column expansion in `deep-atmosphere` configurations.
 """
 function solve_sw!(
-    (; context, fluxb, flux, bcs, op)::NoScatSWRTE,
+    (; context, fluxb, flux, bcs, op, state_cache)::NoScatSWRTE,
     as::AtmosphericState,
     lookup_sw::LookUpSW,
     metric_scaling::M = nothing,
 ) where {M}
-    rte_sw_noscat_solve!(context.device, fluxb, flux, op, bcs, as, lookup_sw)
+    AtmosphericStates.refresh_transposed_state!(state_cache, as)
+    rte_sw_noscat_solve!(
+        context.device,
+        fluxb,
+        flux,
+        op,
+        bcs,
+        as,
+        state_cache,
+        lookup_sw,
+    )
     apply_metric_scaling!(flux, metric_scaling)
 end
 
@@ -201,13 +215,14 @@ Additionally, takes an optional argument `metric_scaling` which scales the resul
 corresponding factor to account for column expansion in `deep-atmosphere` configurations.
 """
 function solve_sw!(
-    (; context, fluxb, flux, band_flux, src, bcs, op)::TwoStreamSWRTE,
+    (; context, fluxb, flux, band_flux, src, bcs, op, state_cache)::TwoStreamSWRTE,
     as::AtmosphericState,
     lookup_sw::LookUpSW,
     lookup_sw_cld::Union{LookUpCld, Nothing} = nothing,
     lookup_sw_aero::Union{LookUpAerosolMerra, Nothing} = nothing,
     metric_scaling::M = nothing,
 ) where {M}
+    AtmosphericStates.refresh_transposed_state!(state_cache, as)
     rte_sw_2stream_solve!(
         context.device,
         fluxb,
@@ -217,6 +232,7 @@ function solve_sw!(
         bcs,
         src,
         as,
+        state_cache,
         lookup_sw,
         lookup_sw_cld,
         lookup_sw_aero,